+++ platforms/safari/FocusView/FocusView/Resources/Script.js
@@ -1,8 +1,8 @@
 function show(enabled, useSettingsInsteadOfPreferences) {
     if (useSettingsInsteadOfPreferences) {
-        document.getElementsByClassName('state-on')[0].innerText = "FocusView’s extension is currently on. You can turn it off in the Extensions section of Safari Settings.";
-        document.getElementsByClassName('state-off')[0].innerText = "FocusView’s extension is currently off. You can turn it on in the Extensions section of Safari Settings.";
-        document.getElementsByClassName('state-unknown')[0].innerText = "You can turn on FocusView’s extension in the Extensions section of Safari Settings.";
+        document.getElementsByClassName('state-on')[0].innerText = "FocusView is on. You can turn it off in Safari Settings > Extensions.";
+        document.getElementsByClassName('state-off')[0].innerText = "FocusView is off. You can turn it on in Safari Settings > Extensions.";
+        document.getElementsByClassName('state-unknown')[0].innerText = "You can turn on FocusView in Safari Settings > Extensions.";
         document.getElementsByClassName('open-preferences')[0].innerText = "Quit and Open Safari Settings…";
     }